chore: move pollBoxModel into a util file
