+++ packages/runtime/src/runtimes/react/controls/slot.tsx
@@ -1,13 +1,12 @@
 import { ComponentPropsWithoutRef, ElementType, ReactNode, useEffect, useState } from 'react'
 import { SlotControl, SlotControlData } from '../../../controls'
 
-import { BoxModel, getBox } from '../../../state/modules/box-models'
-import deepEqual from '../../../utils/deepEqual'
 import { Element } from '../../../runtimes/react'
 import { getIndexes } from '../../../components/utils/columns'
 import { responsiveStyle } from '../../../components/utils/responsive-style'
 import { useStyle } from '../use-style'
 import { cx } from '@emotion/css'
+import { pollBoxModel } from '../pollBoxModel'
 
 export type SlotControlValue = ReactNode
 
@@ -164,34 +163,4 @@ function SlotPlaceholder({ control }: SlotPlaceholderProps): JSX.Element {
       </svg>
     </div>
   )
-}
-
-function pollBoxModel({
-  element,
-  onBoxModelChange,
-}: {
-  element: Element
-  onBoxModelChange(boxModel: BoxModel | null): void
-}): () => void {
-  let currentBoxModel: BoxModel | null = null
-
-  const handleAnimationFrameRequest = () => {
-    const measuredBoxModel = getBox(element)
-
-    if (!deepEqual(currentBoxModel, measuredBoxModel)) {
-      currentBoxModel = measuredBoxModel
-
-      onBoxModelChange(currentBoxModel)
-    }
-
-    animationFrameHandle = requestAnimationFrame(handleAnimationFrameRequest)
-  }
-
-  let animationFrameHandle = requestAnimationFrame(handleAnimationFrameRequest)
-
-  return () => {
-    cancelAnimationFrame(animationFrameHandle)
-
-    onBoxModelChange(null)
-  }
 }
